Split Header into logic and presentation components
